Refactor NFC reading
