+++ src/app/src/main/res/layout/activity_nfc_login.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".NfcLoginActivity">
+    tools:context=".nfc.NfcLoginActivity">
 
     <RelativeLayout
         android:id="@+id/relativeLayout"
